more cleaning to get organized for sync/async setup
