+++ scrapli/driver/core/__init__.py
@@ -3,8 +3,11 @@ from scrapli.driver.core.arista_eos.async_driver import AsyncEOSDriver
 from scrapli.driver.core.arista_eos.driver import EOSDriver
 from scrapli.driver.core.cisco_iosxe.async_driver import AsyncIOSXEDriver
 from scrapli.driver.core.cisco_iosxe.driver import IOSXEDriver
+from scrapli.driver.core.cisco_iosxr.async_driver import AsyncIOSXRDriver
 from scrapli.driver.core.cisco_iosxr.driver import IOSXRDriver
+from scrapli.driver.core.cisco_nxos.async_driver import AsyncNXOSDriver
 from scrapli.driver.core.cisco_nxos.driver import NXOSDriver
+from scrapli.driver.core.juniper_junos.async_driver import AsyncJunosDriver
 from scrapli.driver.core.juniper_junos.driver import JunosDriver
 
 __all__ = (
@@ -12,7 +15,10 @@ __all__ = (
     "EOSDriver",
     "AsyncIOSXEDriver",
     "IOSXEDriver",
+    "AsyncIOSXRDriver",
     "IOSXRDriver",
+    "AsyncNXOSDriver",
     "NXOSDriver",
+    "AsyncJunosDriver",
     "JunosDriver",
 )
