+++ scrapli/driver/core/cisco_iosxe/base_driver.py
@@ -1,0 +1,38 @@
+"""scrapli.driver.core.cisco_iosxe.base_driver"""
+from scrapli.driver.base_network_driver import PrivilegeLevel
+
+PRIVS = {
+    "exec": (
+        PrivilegeLevel(
+            pattern=r"^[a-z0-9.\-@()/:]{1,32}>$",
+            name="exec",
+            previous_priv="",
+            deescalate="",
+            escalate="",
+            escalate_auth=False,
+            escalate_prompt="",
+        )
+    ),
+    "privilege_exec": (
+        PrivilegeLevel(
+            pattern=r"^[a-z0-9.\-@/:]{1,32}#$",
+            name="privilege_exec",
+            previous_priv="exec",
+            deescalate="disable",
+            escalate="enable",
+            escalate_auth=True,
+            escalate_prompt=r"^[pP]assword:\s?$",
+        )
+    ),
+    "configuration": (
+        PrivilegeLevel(
+            pattern=r"^[a-z0-9.\-@/:]{1,32}\(conf[a-z0-9.\-@/:]{0,32}\)#$",
+            name="configuration",
+            previous_priv="privilege_exec",
+            deescalate="end",
+            escalate="configure terminal",
+            escalate_auth=False,
+            escalate_prompt="",
+        )
+    ),
+}
