+++ scrapli/driver/core/cisco_iosxr/async_driver.py
@@ -4,7 +4,7 @@ from typing import Any, Callable, Dict, List, Optional
 
 from scrapli.driver import AsyncNetworkDriver
 from scrapli.driver.base_network_driver import PrivilegeLevel
-from scrapli.driver.core.cisco_iosxr.driver import PRIVS
+from scrapli.driver.core.cisco_iosxr.base_driver import PRIVS
 
 
 async def iosxr_on_open(conn: AsyncNetworkDriver) -> None:
