+++ scrapli/driver/core/cisco_iosxr/driver.py
@@ -4,42 +4,7 @@ from typing import Any, Callable, Dict, List, Optional
 
 from scrapli.driver import NetworkDriver
 from scrapli.driver.base_network_driver import PrivilegeLevel
-
-PRIVS = {
-    "privilege_exec": (
-        PrivilegeLevel(
-            pattern=r"^[a-z0-9.\-@/:]{1,32}#\s?$",
-            name="privilege_exec",
-            previous_priv="",
-            deescalate="",
-            escalate="",
-            escalate_auth=False,
-            escalate_prompt="",
-        )
-    ),
-    "configuration": (
-        PrivilegeLevel(
-            pattern=r"^[a-z0-9.\-@/:]{1,32}\(config[a-z0-9.\-@/:]{0,32}\)#\s?$",
-            name="configuration",
-            previous_priv="privilege_exec",
-            deescalate="end",
-            escalate="configure terminal",
-            escalate_auth=False,
-            escalate_prompt="",
-        )
-    ),
-    "configuration_exclusive": (
-        PrivilegeLevel(
-            pattern=r"^[a-z0-9.\-@/:]{1,32}\(config[a-z0-9.\-@/:]{0,32}\)#\s?$",
-            name="configuration_exclusive",
-            previous_priv="privilege_exec",
-            deescalate="end",
-            escalate="configure exclusive",
-            escalate_auth=False,
-            escalate_prompt="",
-        )
-    ),
-}
+from scrapli.driver.core.cisco_iosxr.base_driver import PRIVS
 
 
 def iosxr_on_open(conn: NetworkDriver) -> None:
