+++ scrapli/driver/core/cisco_nxos/async_driver.py
@@ -3,7 +3,7 @@ from typing import Any, Callable, Dict, List, Optional
 
 from scrapli.driver import AsyncNetworkDriver
 from scrapli.driver.base_network_driver import PrivilegeLevel
-from scrapli.driver.core.cisco_nxos.driver import PRIVS, NXOSDriverBase
+from scrapli.driver.core.cisco_nxos.base_driver import PRIVS, NXOSDriverBase
 
 
 async def nxos_on_open(conn: AsyncNetworkDriver) -> None:
